+++ pyccel/ast/class_defs.py
@@ -6,12 +6,13 @@
 This module contains all types which define a python class which is automatically recognised by pyccel
 """
 
-from pyccel.ast.builtin_methods.set_methods  import (SetAdd, SetClear, SetCopy, SetPop, SetRemove,
+from pyccel.ast.builtin_methods.set_methods  import (SetAdd, SetClear, SetCopy, SetPop,
                                                      SetDiscard, SetUpdate, SetUnion)
 from pyccel.ast.builtin_methods.list_methods import (ListAppend, ListInsert, ListPop,
                                                      ListClear, ListExtend, ListRemove,
                                                      ListCopy, ListSort)
-from pyccel.ast.builtin_methods.dict_methods  import DictPop, DictPopitem, DictGet
+from pyccel.ast.builtin_methods.dict_methods  import( DictPop, DictPopitem, DictGet,
+                                                     DictSetDefault)
 
 from .builtins   import PythonImag, PythonReal, PythonConjugate
 from .core       import ClassDef, PyccelFunctionDef
@@ -163,7 +164,7 @@ SetClass = ClassDef('set',
             PyccelFunctionDef('copy', func_class = SetCopy),
             PyccelFunctionDef('discard', func_class = SetDiscard),
             PyccelFunctionDef('pop', func_class = SetPop),
-            PyccelFunctionDef('remove', func_class = SetRemove),
+            PyccelFunctionDef('remove', func_class = SetDiscard),
             PyccelFunctionDef('union', func_class = SetUnion),
             PyccelFunctionDef('update', func_class = SetUpdate),
             PyccelFunctionDef('__or__', func_class = SetUnion),
@@ -177,6 +178,7 @@ DictClass = ClassDef('dict',
             PyccelFunctionDef('get', func_class = DictGet),
             PyccelFunctionDef('pop', func_class = DictPop),
             PyccelFunctionDef('popitem', func_class = DictPopitem),
+            PyccelFunctionDef('setdefault', func_class = DictSetDefault),
         ])
 
 #=======================================================================================